last fix

+++ app/src/main/java/com/dsvag/yandex/models/yandex/chart/response/Candles.kt
@@ -2,9 +2,10 @@ package com.dsvag.yandex.models.yandex.chart.response
 
 import com.squareup.moshi.Json
 import com.squareup.moshi.JsonClass
+import java.math.BigDecimal
 
 @JsonClass(generateAdapter = true)
 data class Candles(
     @Json(name = "seriesBefore")
-    val seriesBefore: List<Pair<Long, Double>>,
+    val seriesBefore: List<Pair<Long, BigDecimal>>,
 )
